# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBP Digital Agreement System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateAgreementRequest {
  nama_owner: string;
  nik: string;
  alamat_owner: string;
  jenis_properti: string;
  luas_tanah: string;
  legalitas: string;
  alamat_properti: string;
  harga_nett: string;
  tanggal_perjanjian: string;
}

export interface SignAgreementRequest {
  /** Base64 encoded PNG of the signature */
  signature_data: string;
}

export type AgreementStatus =
  (typeof AgreementStatus)[keyof typeof AgreementStatus];

export const AgreementStatus = {
  draft: "draft",
  signed: "signed",
} as const;

export interface Agreement {
  id: string;
  nomor: string;
  nama_owner: string;
  nik: string;
  alamat_owner: string;
  jenis_properti: string;
  luas_tanah: string;
  legalitas: string;
  alamat_properti: string;
  harga_nett: string;
  tanggal_perjanjian: string;
  signature_url?: string | null;
  pdf_url?: string | null;
  status: AgreementStatus;
  created_at: string;
}