# lib/api-zod/src/generated/types/createAgreementRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBP Digital Agreement System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAgreementRequest {
  nama_owner: string;
  nik: string;
  alamat_owner: string;
  jenis_properti: string;
  luas_tanah: string;
  legalitas: string;
  alamat_properti: string;
  harga_nett: string;
  tanggal_perjanjian: string;
}