# lib/api-zod/src/generated/types/signAgreementRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBP Digital Agreement System API
 * OpenAPI spec version: 0.1.0
 */

export interface SignAgreementRequest {
  /** Base64 encoded PNG of the signature */
  signature_data: string;
}